+++ views/tags/_form.php
@@ -76,15 +76,6 @@ use wdmg\widgets\AliasInput;
                 </div>
             </div>
         </div>
-
-        <?= $form->field($model, 'locale')->widget(SelectInput::class, [
-            'items' => $languagesList,
-            'options' => [
-                'id' => 'news-form-locale',
-                'class' => 'form-control'
-            ]
-        ])->label(Yii::t('app/modules/blog', 'Language')); ?>
-
         <hr/>
         <div class="form-group">
             <?= Html::a(Yii::t('app/modules/blog', '&larr; Back to list'), ['tags/index'], ['class' => 'btn btn-default pull-left']) ?>&nbsp;