+++ views/tags/update.php
@@ -25,7 +25,6 @@ $this->params['breadcrumbs'][] = Yii::t('app/modules/blog', 'Updating');
         <?= $this->render('_form', [
             'module' => $module,
             'model' => $model,
-            'languagesList' => $model->getLanguagesList(false),
         ]); ?>
     </div>
 <?php else: ?>