+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Changelog
 =========
 
 ## 1.2.*
- * Delete tags/categories (clearing taxonomy)
+ * Languages list fixed
+ * Deleting tags/categories (clearing taxonomy)
  * Multilangs and default category fixed
  * Lang attribute for HTML tags and inputs